Use dedicated package for graph related entities
