+++ features/graph/dao/api/src/main/java/org/opennms/netmgt/graph/GraphContainerEntity.java
@@ -26,7 +26,7 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.netmgt.topology;
+package org.opennms.netmgt.graph;
 
 import java.util.ArrayList;
 import java.util.List;
