+++ features/graph/dao/impl/src/test/java/org/opennms/netmgt/graph/dao/hibernate/GraphContainerDaoIT.java
@@ -39,11 +39,11 @@ import org.opennms.core.test.db.annotations.JUnitTemporaryDatabase;
 import org.opennms.netmgt.dao.api.GenericPersistenceAccessor;
 import org.opennms.netmgt.graph.dao.api.EntityProperties;
 import org.opennms.netmgt.graph.dao.api.GraphContainerDao;
-import org.opennms.netmgt.topology.EdgeEntity;
-import org.opennms.netmgt.topology.FocusEntity;
-import org.opennms.netmgt.topology.GraphContainerEntity;
-import org.opennms.netmgt.topology.GraphEntity;
-import org.opennms.netmgt.topology.VertexEntity;
+import org.opennms.netmgt.graph.EdgeEntity;
+import org.opennms.netmgt.graph.FocusEntity;
+import org.opennms.netmgt.graph.GraphContainerEntity;
+import org.opennms.netmgt.graph.GraphEntity;
+import org.opennms.netmgt.graph.VertexEntity;
 import org.opennms.test.JUnitConfigurationEnvironment;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.test.context.ContextConfiguration;
