+++ opennms-dao/src/main/resources/META-INF/opennms/applicationContext-shared.xml
@@ -42,6 +42,7 @@
                 <value>org.opennms.netmgt.flows</value>
                 <value>org.opennms.netmgt.enlinkd</value>
                 <value>org.opennms.netmgt.endpoints</value>
+                <value>org.opennms.netmgt.graph</value>
             </list>
         </property>
         <!--  property name="schemaUpdate" value="true">
